Fixes the "Thinking" command
This is what happens when the interns program, you see, we were doing a
c = random(5) and then when c == 3 we were doing the think command. The
problem is random will return a float value between 0 and 5 which means
that it would have taken an act of god for c to actually equal 3.